chore: Extract navigate back to a single Paper component
