+++ assets/js/pages/DesignPage/index.tsx
@@ -1,11 +1,8 @@
 import * as React from "react";
 import * as Pages from "@/components/Pages";
 import * as Paper from "@/components/PaperContainer";
-import * as Icons from "@tabler/icons-react";
 
 import { Logo } from "@/layouts/DefaultLayout/Logo";
-import { Link } from "@/components/Link";
-
 import { Colors } from "./Colors";
 import { Links } from "./Links";
 import { Buttons } from "./Buttons";
@@ -20,7 +17,7 @@ export function Page() {
   return (
     <Pages.Page title={"Design System"}>
       <Paper.Root size="large">
-        <BackToLobbyLink />
+        <Paper.NavigateBack to={"/"} title="Back to the Lobby" />
         <Paper.Body>
           <TitleRow />
           <Colors />
@@ -33,17 +30,6 @@ export function Page() {
         </Paper.Body>
       </Paper.Root>
     </Pages.Page>
-  );
-}
-
-function BackToLobbyLink() {
-  return (
-    <div className="flex items-center mb-4 mt-12">
-      <Link to={"/"}>
-        <Icons.IconArrowLeft className="text-content-dimmed inline mr-2" size={16} />
-        Back to the Lobby
-      </Link>
-    </div>
   );
 }
