+++ assets/js/pages/ProjectAddPage/page.tsx
@@ -5,7 +5,6 @@ import * as Icons from "@tabler/icons-react";
 import * as Projects from "@/models/projects";
 
 import { useLoadedData } from "./loader";
-import { DimmedLink } from "@/components/Link";
 import { Paths, compareIds } from "@/routes/paths";
 import { useMe } from "@/contexts/CurrentUserContext";
 import { useNavigate } from "react-router-dom";
@@ -44,17 +43,9 @@ function Navigation() {
   if (spaceID && space) {
     const spaceProjectsPath = Paths.spaceProjectsPath(spaceID!);
 
-    return (
-      <div className="flex items-center justify-center mb-4 gap-4">
-        <DimmedLink to={spaceProjectsPath}>Back to {space!.name} Space</DimmedLink>
-      </div>
-    );
+    return <Paper.NavigateBack to={spaceProjectsPath} title={`Back to ${space!.name} Space`} />;
   } else {
-    return (
-      <div className="flex items-center justify-center mb-4 gap-4">
-        <DimmedLink to={Paths.projectsPath()}>Back to Projects</DimmedLink>
-      </div>
-    );
+    return <Paper.NavigateBack to={Paths.projectsPath()} title="Back to Projects" />;
   }
 }
